[GOHW-3]: added loms_db create_order

+++ loms/migrations/20230305232822_create_table_orders.sql
@@ -1,7 +1,7 @@
 -- +goose Up
 -- +goose StatementBegin
 CREATE TABLE IF NOT EXISTS orders (
-    id int8 PRIMARY KEY,
+    id serial PRIMARY KEY,
     user_id int8,
     status order_status
 );
